Non-working version of tests
